refactor: extract time-machine feature folder
Move time-machine module into src/app/features/time-machine/
with screen extracted from the route file.
